+++ src/Equinor.Procosys.Preservation.WebApi/Middleware/AddRoleDocumentation.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace Equinor.Procosys.Preservation.WebApi.Misc
+namespace Equinor.Procosys.Preservation.WebApi.Middleware
 {
     public class AddRoleDocumentation : IOperationFilter
     {
